Fixes #1484 - Rewrite TimePicker in typescript
